test: cast vehicle count to long for accurate comparison in SelectFromTest

+++ tck/src/main/java/ee/jakarta/tck/nosql/query/SelectFromTest.java
@@ -173,7 +173,7 @@ class SelectFromTest extends AbstractTemplateTest {
             Assertions.assertThat(result)
                     .isNotEmpty()
                     .get()
-                    .isEqualTo(vehicles.size());
+                    .isEqualTo((long) vehicles.size());
         } catch (UnsupportedOperationException exp) {
             Assertions.assertThat(exp).isInstanceOf(UnsupportedOperationException.class);
         }
